pyproject.toml
[project]
name = "autogen-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.4.9.2",
    "autogen-core>=0.4.9.2",
    "autogen-ext[mcp,openai]>=0.4.9.2",
    "python-dotenv>=1.0.1",
]
